requirements.txt
google-generativeai==0.8.5
dotenv==0.9.9
faiss_cpu==1.12.0
Pillow==11.3.0
cohere==5.17.0
elasticsearch==8.19.0
fastapi==0.104.1
uvicorn==0.24.0
open-clip-torch==2.16.0
transformers==4.55.2
sentencepiece==0.2.1
accelerate==1.10.0
torchscale==0.3.0
icrawler
pymilvus
python-multipart